+++ DarkCodex/DEBUG.cs
@@ -11,6 +11,7 @@ using Kingmaker.EntitySystem.Entities;
 using Kingmaker.Items;
 using Kingmaker.Items.Slots;
 using Kingmaker.Kingdom;
+using Kingmaker.RuleSystem.Rules;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Buffs;
 using Kingmaker.UnitLogic.Class.Kineticist;
@@ -31,30 +32,6 @@ namespace DarkCodex
 {
     public class DEBUG
     {
-        [HarmonyPatch(typeof(Unrecruit), nameof(Unrecruit.RunAction))]
-        public class Quest1
-        {
-            public static bool Prefix()
-            {
-                return !Settings.StateManager.State.debug_block_unrecruit;
-            }
-        }
-
-        [HarmonyPatch(typeof(EtudesSystem), nameof(EtudesSystem.StartEtude))]
-        public class Quest2
-        {
-            public static bool Prefix(BlueprintEtude bp)
-            {
-                foreach (var str in Settings.StateManager.State.debug_block_quest)
-                {
-                    if (Guid.TryParse(str, out Guid guid))
-                        if (bp.AssetGuid == new BlueprintGuid(guid))
-                            return false;
-                }
-                return true;
-            }
-        }
-
         public class Date //#278
         {
             public static void SetDate()