+++ DarkCodex/Mythic.cs
@@ -1,10 +1,13 @@
 ﻿using DarkCodex.Components;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
+using Kingmaker.Enums;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Abilities.Components;
 using Kingmaker.UnitLogic.Abilities.Components.TargetCheckers;
 using Kingmaker.UnitLogic.ActivatableAbilities;
+using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Mechanics.Components;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -17,38 +20,19 @@ namespace DarkCodex
     {
         public static void createLimitlessBardicPerformance()
         {
-            Resource.Cache.Ensure();
-            var bardic = Helper.ToRef<BlueprintAbilityResourceReference>("e190ba276831b5c4fa28737e5e49e6a6");
+            var bardic_resource = BlueprintGuid.Parse("e190ba276831b5c4fa28737e5e49e6a6");
+            var bardic_prereq = Helper.ToRef<BlueprintFeatureReference>("019ada4530c41274a885dfaa0fbf6218");
 
             var limitless = Helper.CreateBlueprintFeature(
                 "LimitlessBardicPerformance",
                 "Limitless Bardic Performance",
                 "Your inspiration knows no bounds.\nBenefit: You no longer have a limited amount of Bardic Performance rounds per day.",
                 group: FeatureGroup.MythicAbility
+                ).SetComponents(
+                Helper.CreatePrerequisiteFeature(bardic_prereq)
                 ); // todo: icon?
 
-            foreach (var ability in Resource.Cache.Ability)
-            {
-                var logic = ability.GetComponent<AbilityResourceLogic>();
-
-                if (logic == null || logic.m_RequiredResource?.deserializedGuid != bardic.deserializedGuid)
-                    continue;
-
-                Helper.PrintDebug("Patching Mythic Barde " + ability.name);
-                logic.ResourceCostDecreasingFacts.Add(limitless.ToRef2());
-            }
-
-            foreach (var ability in Resource.Cache.Activatable)
-            {
-                var logic = ability.GetComponent<ActivatableAbilityResourceLogic>();
-                if (logic == null || logic.m_RequiredResource?.deserializedGuid != bardic.deserializedGuid)
-                    continue;
-
-                Helper.PrintDebug("Patching Mythic Barde " + ability.name);
-                if (logic.m_FreeBlueprint != null && !logic.m_FreeBlueprint.IsEmpty())
-                    Helper.Print($"ERROR: {ability.name} has already a FreeBlueprint");
-                logic.m_FreeBlueprint = limitless.ToRef2();
-            }
+            setResourceDecreasing(bardic_resource, limitless.ToRef2());
 
             Helper.AddMythicTalent(limitless);
         }
@@ -62,7 +46,7 @@ namespace DarkCodex
                 "Limitless Witch Hexes",
                 "Your curse knows no bounds.\nBenefit: You can use your hexes with no time restriction.",
                 group: FeatureGroup.MythicAbility
-                ); // todo: icon?
+                ); // todo: icon? class requirement
 
             // cooldown based
             foreach (var ability in Resource.Cache.Ability)
@@ -74,7 +58,7 @@ namespace DarkCodex
                 if (check == null)
                     continue;
 
-                Helper.PrintDebug("Patching Mythic Witch " + ability.name);
+                Helper.PrintDebug(" -remove cooldown " + ability.name);
                 var checknew = new AbilityTargetHasFactExcept();
                 checknew.m_CheckedFacts = check.m_CheckedFacts;
                 checknew.Inverted = check.Inverted;
@@ -89,7 +73,7 @@ namespace DarkCodex
 
             Helper.AddMythicTalent(limitless);
         }
-    
+
         public static void createLimitlessSmite()
         {
             var smite_evil = ResourcesLibrary.TryGetBlueprint<BlueprintAbility>("7bb9eb2042e67bf489ccd1374423cdec");
@@ -114,6 +98,131 @@ namespace DarkCodex
             Helper.AddMythicTalent(limitless);
         }
 
+        public static void createLimitlessBombs()
+        {
+            var bomb_resource = BlueprintGuid.Parse("1633025edc9d53f4691481b48248edd7");
+            var bomb_prereq = Helper.ToRef<BlueprintFeatureReference>("54c57ce67fa1d9044b1b3edc459e05e2"); //AlchemistBombsFeature
+
+            var limitless = Helper.CreateBlueprintFeature(
+                "LimitlessBombs",
+                "Limitless Alchemist Bombs",
+                "You learn how to create a philosopher’s stone that turns everything into explosives.\nBenefit: You no longer have a limited amount of Bombs per day.",
+                group: FeatureGroup.MythicAbility
+                ).SetComponents(
+                Helper.CreatePrerequisiteFeature(bomb_prereq)
+                ); // todo: icon?
+
+            setResourceDecreasing(bomb_resource, limitless.ToRef2());
+
+            Helper.AddMythicTalent(limitless);
+        }
+
+        public static void createLimitlessArcanePool()
+        {
+            var arcane_resource = BlueprintGuid.Parse("effc3e386331f864e9e06d19dc218b37");
+            var eldritch_resource = BlueprintGuid.Parse("effc3e386331f864e9e06d19dc218b37");
+            var arcane_prereq = Helper.ToRef<BlueprintFeatureReference>("3ce9bb90749c21249adc639031d5eed1"); //ArcanePoolFeature
+            var eldritch_prereq = Helper.ToRef<BlueprintFeatureReference>("95e04a9e86aa9e64dad7122625b79c62"); //EldritchPoolFeature
+            var abundant_arcane = Helper.ToRef<BlueprintFeatureReference>("8acebba92ada26043873cae5b92cef7b"); //AbundantArcanePool
+
+            var limitless = Helper.CreateBlueprintFeature(
+                "LimitlessArcanePool",
+                "Limitless Arcane Pool",
+                "Benefit: You can use your arcana without using your arcane pool. You still need to spend arcane points for spell recall.",
+                group: FeatureGroup.MythicAbility,
+                icon: abundant_arcane.Get().Icon
+                ).SetComponents(
+                Helper.CreatePrerequisiteFeature(abundant_arcane)
+                );
+
+            setResourceDecreasing(arcane_resource, limitless.ToRef2(), true);
+
+            Helper.AddMythicTalent(limitless);
+        }
+
+        public static void createLimitlessArcaneReservoir()
+        {
+            var arcane_resource = BlueprintGuid.Parse("cac948cbbe79b55459459dd6a8fe44ce");
+            var arcane_prereq = Helper.ToRef<BlueprintFeatureReference>("55db1859bd72fd04f9bd3fe1f10e4cbb"); //ArcanistArcaneReservoirFeature
+            var enforcerer_prereq = Helper.ToRef<BlueprintFeatureReference>("9d1e2212594cf47438fff2fa3477b954"); //ArcaneEnforcerArcaneReservoirFeature
+
+            var limitless = Helper.CreateBlueprintFeature(
+                "LimitlessArcaneReservoir",
+                "Limitless Arcane Reservoir",
+                "Benefit: You can use your arcane exploits without using your arcane reservoir. This does not apply to Magical Supremacy.",
+                group: FeatureGroup.MythicAbility
+                ).SetComponents(
+                Helper.CreatePrerequisiteFeature(arcane_prereq, true),
+                Helper.CreatePrerequisiteFeature(enforcerer_prereq, true)
+                ); // todo: icon?
+
+            setResourceDecreasing(arcane_resource, limitless.ToRef2(), true);
+
+            Helper.AddMythicTalent(limitless);
+        }
+
+        public static void createLimitlessKi()
+        {
+            var ki_resource = BlueprintGuid.Parse("9d9c90a9a1f52d04799294bf91c80a82");
+            var scaled_resource = BlueprintGuid.Parse("7d002c1025fbfe2458f1509bf7a89ce1");
+            var abundant_ki = Helper.ToRef<BlueprintFeatureReference>("e8752f9126d986748b10d0bdac693264"); //AbundantKiPool
+
+            var limitless = Helper.CreateBlueprintFeature(
+                "LimitlessKi",
+                "Limitless Ki",
+                "Your body and mind are in perfect equilibrium.\nBenefit: Your abilities cost one less ki.",
+                group: FeatureGroup.MythicAbility,
+                icon: abundant_ki.Get().Icon
+                ).SetComponents(
+                Helper.CreatePrerequisiteFeature(abundant_ki)
+                );
+
+            setResourceDecreasing(ki_resource, limitless.ToRef2());
+            setResourceDecreasing(scaled_resource, limitless.ToRef2());
+
+            Helper.AddMythicTalent(limitless);
+        }
+
+        public static void createLimitlessDomain()
+        {
+            var limitless = Helper.CreateBlueprintFeature(
+                "LimitlessDomainPowers",
+                "Limitless Domain Powers",
+                "You are chosen by your deity.\nBenefit: You can use the abilities of your domains at will.",
+                group: FeatureGroup.MythicAbility
+                );
+
+            foreach (var ability in Resource.Cache.Ability)
+            {
+                var logic = ability.GetComponent<AbilityResourceLogic>();
+                if (logic == null)
+                    continue;
+                if (logic.CostIsCustom)
+                    continue;
+                if (!ability.name.Contains("Domain"))
+                    continue;
+
+                logic.ResourceCostDecreasingFacts.Add(limitless.ToRef2());
+                Helper.PrintDebug(" -free ability " + ability.name);
+            }
+
+            foreach (var ability in Resource.Cache.Activatable)
+            {
+                var logic = ability.GetComponent<ActivatableAbilityResourceLogic>();
+                if (logic == null)
+                    continue;
+                if (!ability.name.Contains("Domain"))
+                    continue;
+
+                if (logic.m_FreeBlueprint != null && !logic.m_FreeBlueprint.IsEmpty())
+                    Helper.Print($"ERROR: {ability.name} has already a FreeBlueprint");
+                logic.m_FreeBlueprint = limitless.ToRef2();
+                Helper.PrintDebug(" -free activatable " + ability.name);
+            }
+
+            Helper.AddMythicTalent(limitless);
+        }
+
         public static void createKineticMastery()
         {
             var kineticist_class = Helper.ToRef<BlueprintCharacterClassReference>("42a455d9ec1ad924d889272429eb8391");
@@ -121,7 +230,7 @@ namespace DarkCodex
             var kinetic_mastery = Helper.CreateBlueprintFeature(
                 "KineticMastery",
                 "Kinetic Mastery",
-                "You mastered the elements. Benefit: You add your mythic rank to physical kinetic blasts and half your mythic rank to energy kinetic blasts.",
+                "You mastered the elements. Benefit: You add your mythic rank to attack rolls of physical kinetic blasts and half your mythic rank to attack rolls of energy kinetic blasts.",
                 group: FeatureGroup.MythicFeat
                 ).SetComponents(
                 Helper.CreatePrerequisiteClassLevel(kineticist_class, 1),
@@ -130,5 +239,53 @@ namespace DarkCodex
 
             Helper.AddMythicFeat(kinetic_mastery);
         }
+
+        public static void patchKineticOvercharge()
+        {
+            var overcharge = ResourcesLibrary.TryGetBlueprint<BlueprintFeature>("4236ca275c6fa9b4a9945b2645262616");
+            overcharge.m_Description = Helper.CreateString("You use your mythic powers to fuel kineticist abilities.\nBenefit: Reduce the burn cost of kineticist blasts by one burn point.");
+            overcharge.Ranks = 6;
+            overcharge.RemoveComponents(default(AddFacts));
+            overcharge.AddComponents(
+                new KineticistReduceBurnPooled() { ReduceBurn = Helper.CreateContextValue(AbilityRankType.Default) },
+                Helper.CreateContextRankConfig(ContextRankBaseValueType.FeatureRank, feature: overcharge.ToRef())
+                );
+        }
+
+        #region Helper
+
+        public static void setResourceDecreasing(BlueprintGuid resource, BlueprintUnitFactReference limitless, bool repeat = false)
+        {
+            Resource.Cache.Ensure();
+
+            foreach (var ability in Resource.Cache.Ability)
+            {
+                var logic = ability.GetComponent<AbilityResourceLogic>();
+                if (logic == null || logic.m_RequiredResource?.deserializedGuid != resource)
+                    continue;
+                if (logic.CostIsCustom)
+                    continue;
+
+                int total = !repeat ? 1 : logic.Amount + logic.ResourceCostIncreasingFacts.Count;
+                for (int i = 0; i < total; i++)
+                    logic.ResourceCostDecreasingFacts.Add(limitless);
+                Helper.PrintDebug(" -free ability " + ability.name);
+            }
+
+            foreach (var ability in Resource.Cache.Activatable)
+            {
+                var logic = ability.GetComponent<ActivatableAbilityResourceLogic>();
+                if (logic == null || logic.m_RequiredResource?.deserializedGuid != resource)
+                    continue;
+
+                Helper.PrintDebug(" -free activatable " + ability.name);
+                if (logic.m_FreeBlueprint != null && !logic.m_FreeBlueprint.IsEmpty())
+                    Helper.Print($"ERROR: {ability.name} has already a FreeBlueprint");
+                logic.m_FreeBlueprint = limitless;
+            }
+        }
+        
+        #endregion
+
     }
 }